feat(kjd): shared page aggregate
