+++ apps/kjd/app/blog/page.tsx
@@ -1,9 +1,9 @@
 import {
   BlogPage,
-  pageMetadataGenerator,
+  blogPageMetadataGenerator,
 } from '@nxify/kjd-feature-blog/server';
 
-export const generateMetadata = pageMetadataGenerator;
+export const generateMetadata = blogPageMetadataGenerator;
 export const revalidate = 10;
 
 export default BlogPage;
