+++ libs/kjd/ui-page-fragments/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/server/article-content/article-content';
 // Use this file to export React client components (e.g. those with 'use client' directive) or other non-server utilities
 export * from './server';
+export * from './lib/server/page/page';
